backend/requirements.txt
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-language==2.0.0
google-cloud-speech==2.3.0
google-cloud-texttospeech==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.18
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.17.0
python-engineio==4.1.0
python-socketio==5.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
twilio==6.56.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==1.0.1
